backend/requirements.txt
asgiref==3.8.1
boto3==1.39.3
botocore==1.39.3
certifi==2025.6.15
charset-normalizer==3.4.2
cloudinary==1.44.1
Django==5.2.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.4
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
stripe==12.2.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
